+++ blog-next/src/App.tsx
@@ -7,23 +7,29 @@ import NewArticle from "./pages/NewArticle"
 import Article from "./pages/Article"
 import Profile from "./pages/Profile"
 import Message from "./pages/Message"
+import { ToastProvider } from "./components/Toast"
+import { ConfirmDialogProvider } from "./components/ConfirmDialog"
 
 function App() {
   return (
-    <Router>
-      <Navbar />
-      <div className="container mx-auto px-4 py-8">
-        <Routes>
-          <Route path="/" element={<Home />} />
-          <Route path="/login" element={<Login />} />
-          <Route path="/register" element={<Register />} />
-          <Route path="/new" element={<NewArticle />} />
-          <Route path="/article/:id" element={<Article />} />
-          <Route path="/profile" element={<Profile />} />
-          <Route path="/message" element={<Message />} />
-        </Routes>
-      </div>
-    </Router>
+    <ToastProvider>
+      <ConfirmDialogProvider>
+        <Router>
+          <Navbar />
+          <div className="container mx-auto px-4 py-8">
+            <Routes>
+              <Route path="/" element={<Home />} />
+              <Route path="/login" element={<Login />} />
+              <Route path="/register" element={<Register />} />
+              <Route path="/new" element={<NewArticle />} />
+              <Route path="/article/:id" element={<Article />} />
+              <Route path="/profile" element={<Profile />} />
+              <Route path="/message" element={<Message />} />
+            </Routes>
+          </div>
+        </Router>
+      </ConfirmDialogProvider>
+    </ToastProvider>
   )
 }